* Linux névtér létrehozva.
