+++ GrabTheMoment/Linux/SpecialKey.cs
@@ -1,0 +1,12 @@
+﻿#if __MonoCS__
+using System;
+
+namespace GrabTheMoment.Linux
+{
+    public enum SpecialKey
+    {
+        // TODO: Print gombra átrakni.
+        Print = Gdk.Key.Page_Up
+    };
+}
+#endif
