+++ GrabTheMoment/Linux/SpecialKeyPressedHandler.cs
@@ -1,0 +1,8 @@
+﻿#if __MonoCS__
+using System;
+
+namespace GrabTheMoment.Linux
+{
+    public delegate void SpecialKeyPressedHandler(object o, SpecialKey key);
+}
+#endif
